Refactor play function and finish parsing
Refactor of play function:
The info dictionary will now be updated in a different function all
together. When calling the function, all information that can be updated
will be, even if they are not needed for the next prompt.
A lot of smaller tasks are now in their own methods. The most important
part is, that the info dict is not used anymore to retrieve information
from, but only as a step to convert information into a readable string.
The important game information are now saved in their own dictionaries:
played_cards, predictions, points, tricks_per_player, playing_order

Finish parsing:
The new prompt_model method is a wrapper function for the whole
prompting, parsing and reprompting process. Depending on the value of
liberal_mode (that can be set in the instancegenerator), the model is
reprompted with a specific text.